Implement auth in User service

+++ src/services/userService.ts
@@ -1,15 +1,17 @@
 'use server'
-import { hashPassword } from '@/utils/auth'
+import { hashString, verifySession } from '@/utils/auth'
 import { randomUUID } from 'crypto'
 import path from 'path'
 import fs from 'fs'
 import { CreateUserDTO, PatchUserDTO, User } from '@/types/User'
+import { Session } from '@/types/Auth'
 
 export const createUser = async (formData: CreateUserDTO): Promise<boolean> => {
     const { login, password, email } = formData
     const userId = randomUUID()
     const folderPath = path.join(process.cwd(), 'src', 'data', 'users', login)
-    const hashedPassowrd = await hashPassword(password)
+    const hashedPassowrd = await hashString(password)
+
     const user: User = {
         id: userId,
         login,
@@ -17,75 +19,75 @@ export const createUser = async (formData: CreateUserDTO): Promise<boolean> => {
         password: hashedPassowrd,
     }
     try {
-        fs.mkdirSync(folderPath, { recursive: true })
+        fs.mkdirSync(folderPath)
     } catch {
         throw new Error('User with login: ' + login + ' already exists')
     }
     const filePath = path.join(folderPath, 'user.json')
     fs.writeFileSync(filePath, JSON.stringify(user, null, 2), 'utf8')
 
-    //add session
-
     return new Promise((resolve) => {
         resolve(true)
     })
 }
 
-export const getUser = async (): Promise<User> => {
-    const userLogin = 'test' //swap this to login from cookies
+export const getUser = async (session: Session): Promise<User> => {
     const filePath = path.join(
         process.cwd(),
         'src',
         'data',
         'users',
-        userLogin,
+        session.login,
         'user.json'
     )
     let data
     try {
         data = fs.readFileSync(filePath, 'utf8')
     } catch {
-        throw new Error('Unexpected error')
+        throw new Error('User with login:' + session.login + ' does not exist')
     }
     const user: User = JSON.parse(data)
-
-    return new Promise((resolve) => {
-        //check is session and user is correct, reject if its inwalid
-        resolve(user)
+    const verification = await verifySession(session)
+    return new Promise((resolve, reject) => {
+        if (verification) {
+            resolve(user)
+        }
+        reject('Session is invalid')
     })
 }
 
-export const deleteUser = async (): Promise<boolean> => {
-    const userLogin = 'test2' //swap this to user id from cookies
-    const folderPath = path.join(process.cwd(), 'src', 'data', 'users', userLogin)
-    try {
-        fs.rmdirSync(folderPath, { recursive: true })
-    } catch {
-        throw new Error('Unexpected error')
-    }
-
-    return new Promise((resolve) => {
-        resolve(true)
+export const deleteUser = async (session: Session): Promise<boolean> => {
+    const folderPath = path.join(process.cwd(), 'src', 'data', 'users', session.login)
+    const verification = await verifySession(session)
+    return new Promise((resolve, reject) => {
+        if (verification) {
+            fs.rmdirSync(folderPath, { recursive: true })
+            resolve(true)
+        }
+        reject('Session is invalid')
     })
 }
 
-export const patchUser = async (formData: PatchUserDTO): Promise<boolean> => {
-    const { login, password, email } = formData
-    const userLogin = 'test' //swap this to login from cookies
+export const patchUser = async (
+    formData: PatchUserDTO,
+    session: Session
+): Promise<boolean> => {
+    const { login, password, email, sessionId } = formData
+
     let filePath = path.join(
         process.cwd(),
         'src',
         'data',
         'users',
-        userLogin,
+        session.login,
         'user.json'
     )
 
-    const user: User = await getUser()
+    const user: User = await getUser(session)
 
-    if (login != undefined) {
+    if (login !== undefined) {
         user.login = login
-        const oldFolderPath = path.join('src', 'data', 'users', userLogin)
+        const oldFolderPath = path.join('src', 'data', 'users', session.login)
 
         const newFolderPath = path.join(process.cwd(), 'src', 'data', 'users', login)
         const newFilePath = path.join(newFolderPath, 'user.json')
@@ -96,22 +98,27 @@ export const patchUser = async (formData: PatchUserDTO): Promise<boolean> => {
             throw new Error('User with login: ' + login + ' already exists')
         }
         filePath = newFilePath
-
-        //change userLogin in session
     }
 
-    if (password != undefined) {
-        const hashedPassowrd = await hashPassword(password)
+    if (password !== undefined) {
+        const hashedPassowrd = await hashString(password)
         user.password = hashedPassowrd
     }
 
-    if (email != undefined) {
+    if (email !== undefined) {
         user.email = email
     }
 
-    fs.writeFileSync(filePath, JSON.stringify(user, null, 2), 'utf8')
+    if (sessionId !== undefined) {
+        user.sessionId = sessionId
+    }
 
-    return new Promise((resolve) => {
-        resolve(true)
+    const verification = await verifySession(session)
+    return new Promise((resolve, reject) => {
+        if (verification) {
+            fs.writeFileSync(filePath, JSON.stringify(user, null, 2), 'utf8')
+            resolve(true)
+        }
+        reject('Session is invalid')
     })
 }
